WIP ei socket for at-spi2-core
This works with the `receive` example of `reis` using
`LIBEI_SOCKET=/tmp/atspi-ei-kb.socket`. I've also worked on changes to
at-spi2-core, so I'll now be able to try testing it...

If a version of this is ultimately used, it will need a secure way to
pass the socket to accessibility tools. Putting it in `/tmp` is a
placeholder.